# Dynamically determine the amount of work that can be performed in an endpoint iteration