@Ralith 2021-05-30
Haven't carefully reviewed the implementation yet, but overall this looks reasonable to me. I'd be happier if we had empirical evidence of an environment where the hardcoded values are significantly bad, but that's somewhat mitigated by the good encapsulation here and the infrequent sampling.