@Matthias247 2021-05-31
Seems like even though I increased the timer and gave it a lot of grace time the test is still flaky on CI. Maybe the CI macos CI runner is rather overscheduled. I will look into in way to mock time for the test to de-flake it.