@Ralith 2021-05-31
Ultimately we want to get rid of these tests since they're a maintenance burden, we just need to investigate making up for any coverage losses that would entail.