@Matthias247 2021-05-31
The last failure was about the newly added test, not about the H3 one. I made this one now deterministic (at the cost of some other ugliness in the implementation)