@Matthias247 2021-06-12
I'm not sure what happened to those macos tests:

```
---- tests::echo_dualstack stdout ----
thread 'tests::echo_dualstack' panicked at 'called `Result::unwrap()` on an `Err` value: Os { code: 22, kind: InvalidInput, message: "Invalid argument" }', quinn/src/tests.rs:597:56
```

However I feel like those failures are unrelated to these changes