@Matthias247 2021-06-13
CI failure was most likely caused by an issue in mio 0.7.12 which was picked up by the build: https://github.com/tokio-rs/mio/issues/1497

That version is now yanked, so retrying the run might lead to success